aulao-1-ia-em-producao/etl-monorepo/requirements.txt
google-cloud-storage>=2.0.0
google-cloud-bigquery>=3.0.0
pandas>=2.0.0
pyarrow>=14.0.0
requests>=2.31.0
sqlalchemy>=2.0.0
pyyaml>=6.0.0
openpyxl>=3.1.0
